backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
marshmallow==3.21.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0

Flask-CORS==3.0.10
